process names for 03_pstree_kill where wrong
the shebang for the corresponding shellscripts ended up being
  #!/usr/bin/env sh
which made the process name (with ps) end up being "sh" rather than the
filename of the script.

I now explicitly (re) set the shebang when copying the scripts to make
sure to use the appropriate shell.
(That should work, even on NixOS, which was the main reason for using
"/usr/bin/env sh" shebang).